requirements.txt
python-dotenv
requests
beautifulsoup4
docopts
python-dotenv
selenium
Selenium-Screenshot
selenium-stealth
pyvirtualdisplay
pymsteams
ghapi
jsonlines
